+++ src/bda/plone/ticketshop/browser/views.py
@@ -1,14 +1,14 @@
 from Acquisition import aq_parent
 from Products.Five import BrowserView
 from bda.plone.cart import ascur
+from bda.plone.cart import get_object_by_uid
 from bda.plone.orders.browser import views
 from bda.plone.shop.at import field_value
 from bda.plone.ticketshop.interfaces import ITicket
 from bda.plone.ticketshop.interfaces import ITicketOccurrence
-from plone.app.uuid.utils import uuidToObject
 from plone.event.interfaces import IEvent
 from plone.app.event.browser.event_summary import EventSummaryView
-
+from zope.component.hooks import getSite
 
 class EventTicketSummaryView(EventSummaryView):
 
@@ -44,7 +44,7 @@ class TicketOccurrenceView(TicketView):
 
 
 def _get_booking_url(booking):
-    booking_obj = uuidToObject(booking.attrs['buyable_uid'])
+    booking_obj = get_object_by_uid(getSite(), booking.attrs['buyable_uid'])
     url = booking_obj.absolute_url()
     if ITicketOccurrence.providedBy(booking_obj):
         # assumption1: ticketoccurrence id = occurrence id